# Ingest performance