--- a/crux-core/src/crux/fork.clj
+++ b/crux-core/src/crux/fork.clj
@@ -120,6 +120,13 @@
                            capped-valid-time
                            capped-tx-id))
 
+  db/IndexMeta
+  (-read-index-meta [this k not-found]
+    (let [v (db/read-index-meta transient-index-snapshot k ::not-found)]
+      (if (not= v ::not-found)
+        v
+        (db/read-index-meta persistent-index-snapshot k not-found))))
+
   java.io.Closeable
   (close [_]
     (cio/try-close transient-index-snapshot)
@@ -160,15 +167,6 @@
   (store-index-meta [this k v]
     (db/store-index-meta transient-index-store k v))
 
-  (read-index-meta [this k]
-    (db/read-index-meta this k nil))
-
-  (read-index-meta [this k not-found]
-    (let [v (db/read-index-meta transient-index-store k ::not-found)]
-      (if (not= v ::not-found)
-        v
-        (db/read-index-meta persistent-index-store k not-found))))
-
   (latest-completed-tx [this]
     (or (db/latest-completed-tx transient-index-store)
         (db/latest-completed-tx persistent-index-store)))
@@ -179,6 +177,13 @@
   (tx-failed? [this tx-id]
     (or (db/tx-failed? transient-index-store tx-id)
         (db/tx-failed? persistent-index-store tx-id)))
+
+  db/IndexMeta
+  (-read-index-meta [this k not-found]
+    (let [v (db/read-index-meta transient-index-store k ::not-found)]
+      (if (not= v ::not-found)
+        v
+        (db/read-index-meta persistent-index-store k not-found))))
 
   (open-index-snapshot ^java.io.Closeable [this]
     (->ForkedIndexSnapshot (or persistent-index-snapshot (db/open-index-snapshot persistent-index-store))