Fixes #1252

Main changes:
* It's now the index store that's responsible for the in-memory transaction, so that it can re-use the encoded documents - previously, we were calling `->content-idx-kvs` and `etx-kvs` twice for all the documents/tx-events respectively. Because we're working at a lower-level, we can pipe the encoded KVs in the in-memory KV store straight into Rocks.
* In-memory index-store uses a `j.u.TreeMap`, rather than `c.l.PersistentTreeMap`. The 'snapshots' yielded from this, strictly speaking, aren't snapshots - they're mutable, but they're only ever read from one thread, and that thread is _either_ reading or writing, never both at once - i.e. it's immutable through the sections that matter.
* We were calling `unindex-eids` even when there weren't any eids to unindex - this didn't have any `seq` checks, and opened/closed snapshots regardless. `(when (seq eids) ...)` cut about 10% of my microbench.
* I've replaced a few usages of `merge` with `into` - the former does `(reduce conj ...)`, which doesn't use transient data structures.
* Stats are now updated synchronously - before/after wasn't conclusive on this one on my microbench, but makes the code simpler.
* I've not removed the check where we read the KV store to decide whether to index the doc, one for another PR.

Microbench:
* Attach profiler
* Submit 25 transactions with 1k simple docs, `crux/sync`
* Restart node, time `crux/sync`

Timings (YourKit, 'tracing' - about 10-15x slower than just running at the REPL, but more interested in the relative speedup):
* Before: 25s
* After TreeMap, unindex-eids, merge -> into changes: 17s
* After moving fork from tx -> index-store: 12s
* Particularly, `InFlightTx.commit` comes down from 5.2s to 0.7s

In bench, we're seeing 35-40% improvements in ts-devices ingest against master, run in progress, ~15-20% down against pre 11/11 - will update with more stats as they come in.